// src/modules/product/dto/product.dto.ts
export type ProductDto = {
  productid: number;
  name: string;
  link: string;
  description: string;
  product_include: string;
  harvest_cycle: string;
  flavour_characteristics: string;
  height: string;
  width: string;
  care_technique: string;
  harvesting_technique: string;
  image_urls: string[];
  slug: string;
  price: number;
  price_id: string;
  category: string;
};